+++ phonemizer/backend/espeak.py
@@ -31,6 +31,12 @@ from phonemizer.logger import get_logger
 _ESPEAK_FLAGS_RE = re.compile(r'\(.+?\)')
 
 
+# a global variable being used to verload the default espeak installed on the
+# system. The user can choose an alternative espeak with the method
+# EspeakBackend.set_espeak_path().
+_ESPEAK_DEFAULT_PATH = None
+
+
 class EspeakBackend(BaseBackend):
     """Espeak backend for the phonemizer"""
 
@@ -40,6 +46,7 @@ class EspeakBackend(BaseBackend):
                  language_switch='keep-flags', with_stress=False,
                  logger=get_logger()):
         super(self.__class__, self).__init__(language, logger=logger)
+        self.logger.debug(f'espeak is {self.espeak_path()}')
 
         # adapt some command line option to the espeak version (for
         # phoneme separation and IPA output)
@@ -76,7 +83,31 @@ class EspeakBackend(BaseBackend):
         return 'espeak'
 
     @staticmethod
-    def espeak_exe():
+    def set_espeak_path(fpath):
+        """"""
+        global _ESPEAK_DEFAULT_PATH
+        if not fpath:
+            _ESPEAK_DEFAULT_PATH = None
+            return
+
+        if not (os.path.isfile(fpath) and os.access(fpath, os.X_OK)):
+            raise ValueError(
+                f'{fpath} is not an executable file')
+
+        _ESPEAK_DEFAULT_PATH = os.path.abspath(fpath)
+
+    @staticmethod
+    def espeak_path():
+        if 'ESPEAK_PATH' in os.environ:
+            espeak = os.environ['ESPEAK_PATH']
+            if not (os.path.isfile(espeak) and os.access(espeak, os.X_OK)):
+                raise ValueError(
+                    f'ESPEAK_PATH={espeak} is not an executable file')
+            return os.path.abspath(espeak)
+
+        if _ESPEAK_DEFAULT_PATH:
+            return _ESPEAK_DEFAULT_PATH
+
         espeak = distutils.spawn.find_executable('espeak-ng')
         if not espeak:  # pragma: nocover
             espeak = distutils.spawn.find_executable('espeak')
@@ -84,12 +115,12 @@ class EspeakBackend(BaseBackend):
 
     @classmethod
     def is_available(cls):
-        return True if cls.espeak_exe() else False
+        return True if cls.espeak_path() else False
 
     @classmethod
     def long_version(cls):
         return subprocess.check_output(shlex.split(
-            '{} --help'.format(cls.espeak_exe()), posix=False)).decode(
+            '{} --help'.format(cls.espeak_path()), posix=False)).decode(
                 'utf8').split('\n')[1]
 
     @classmethod
@@ -104,13 +135,16 @@ class EspeakBackend(BaseBackend):
         long_version = cls.long_version()
 
         # extract the version number with a regular expression
-        return re.match(cls.espeak_version_re, long_version).group(1)
+        try:
+            return re.match(cls.espeak_version_re, long_version).group(1)
+        except AttributeError:
+            raise RuntimeError(f'cannot extract espeak version from {cls.espeak_path()}')
 
     @classmethod
     def supported_languages(cls):
         # retrieve the languages from a call to 'espeak --voices'
         voices = subprocess.check_output(shlex.split(
-            '{} --voices'.format(cls.espeak_exe()), posix=False)).decode(
+            '{} --voices'.format(cls.espeak_path()), posix=False)).decode(
                 'utf8').split('\n')[1:-1]
         voices = [v.split() for v in voices]
 
@@ -158,7 +192,7 @@ class EspeakBackend(BaseBackend):
 
                     # generate the espeak command to run
                     command = '{} -v{} {} -q -f {} {}'.format(
-                        self.espeak_exe(), self.language, self.ipa,
+                        self.espeak_path(), self.language, self.ipa,
                         data.name, self.sep)
 
                     if self.logger: